Visually, the username is now centered relative to the avatar. Move IconButton back to UserName.
