+++ web-client/src/components/ProfilePage/UserName/index.module.css
@@ -11,6 +11,7 @@
   font-weight: normal;
   font-size: 16px;
   line-height: 24px;
+  padding-left: 10px;
   /* identical to box height, or 150% */
 
   /* neutral/10 */
